pyproject.toml
[project]
name = "tiny-sales"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "tortoise-orm>=0.21.3",
    "svix-ksuid>=0.6.2",
    "aiosqlite>=0.20.0",
    "uvicorn[standard]>=0.30.1",
    "aerich[toml]>=0.8.0",
    "python-jose[cryptography]>=3.5.0",
    "bcrypt>=4.3.0",
]

[tool.uv]
package = true

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
manage-users = "app.cli.main:app"

[tool.aerich]
tortoise_orm = "app.main.TORTOISE_ORM_CONFIG"
location = "./migrations"

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-httpx>=0.21.2",
] # httpx is already installed, this adds the pytest plugin
